Leverage usertype
